requirements.txt
# Alvin AI Assistant Requirements
ollama==0.1.0
asyncio==3.4.3
websockets==12.0
pyaudio==0.2.14
torch==2.1.0
Flask==3.0.3
pvporcupine
sounddevice==0.4.6
huggingface_hub==0.23.0
transformers==4.41.2
pytesseract==0.3.10
pyautogui==0.9.54
Pillow==10.3.0
psutil==5.9.8
python-dotenv==1.0.1
GPUtil
